% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_called.R
\name{expect_called}
\alias{expect_called}
\title{Add expectation to a function.}
\usage{
expect_called(func, to_mock, expected_call_list)
}
\arguments{
\item{func}{description}

\item{to_mock}{description}

\item{expected_call_list}{description}
}
\description{
Works with named or unnamed call list. If unnamed, must be
a complete call list, if named can be partial.
}

