% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_source_utils.R
\name{get_function_source}
\alias{get_function_source}
\title{Returns the pure R source code for a function.}
\usage{
get_function_source(func)
}
\arguments{
\item{func}{function for which to retreive the source code.}
}
\description{
Returns the pure R source code for a function.
}

