% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockery.R
\docType{package}
\name{mockery-package}
\alias{mockery}
\alias{mockery-package}
\title{R package to make mocking easier}
\description{
There are great tools for unit testing in R out there already but
they don't come with a lot of support for mock objects. This
package aims at fixing that.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-lib/mockery}
  \item Report bugs at \url{https://github.com/r-lib/mockery/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@posit.co}

Authors:
\itemize{
  \item Noam Finkelstein
  \item Lukasz Bartnik
  \item Jim Hester
}

}
