% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stub.R
\name{stub}
\alias{stub}
\title{Replace a function with its stub.}
\usage{
stub(func, to_stub, stub)
}
\arguments{
\item{func}{description}

\item{to_stub}{description}

\item{stub}{description}
}
\description{
Replace a function with its stub.
}

